Stage 3/4: Saved and loaded Complete
The program should provide a menu with five options:

Generate a new maze.
Load a maze.
Save the maze.
Display the maze.
Exit.
After a maze is generated or loaded from a file, it becomes the current maze that can be saved or displayed.

If there is no current maze (generated or loaded), a user should not see the options save and display the maze. If a user chooses an option that requires a file, he must enter a path to the file. You must always check the result of processing files and display user-friendly messages.

The program should output the maze to the user only in two scenarios:

After generating a maze;
After choosing an option display the maze.
Your program must successfully handle the following cases:

if an incorrect option was chosen, the program must print a message like "Incorrect option. Please try again";
if a file to load a maze does not exist, the program should not stop, it must print a message like "The file ... does not exist";
if a file has an invalid format for a maze, the program should not stop, but it must print a message like "Cannot load the maze. It has an invalid format".
By the way, you can store the maze in any format, the tests do not check the contents of the file. The most important thing is that the saved maze must be correctly loaded into the program.

-----------------------------------------------------------------

I created a adjacency matrix and based on it a minimal spanning tree. Then I created an array of values for walls and passageways for the maze based on the data from this tree. Saving and loading was implemented using serialization of the Maze object.