Work on project. Stage 4/4: A way out : Complete.
I used an Dijkstra algorithm to find a way out of the maze. This site helped you understand how it works : https://www.cs.usfca.edu/~galles/visualization/Dijkstra.html
Although then i had to connect neighboring nodes in the resulting minimal tree in the final array of the maze. ))
Saving and loading is implemented with serialization of the "Maze" object, so the file test_maze.txt so big and not understand.
The resulting code did not clear from debugging messages (thanks to Mariano for idea) so that it should be useful in future projects with minimal spanning tree searches.